refactor: 테스트 클래스 패키지 변경
domain 하위 클래스에 도메인별 테스트가 위치하도록 변경
