+++ src/test/java/com/nexters/keyme/domain/notification/service/NotificationServiceTest.java
@@ -1,4 +1,4 @@
-package com.nexters.keyme.notification.service;
+package com.nexters.keyme.domain.notification.service;
 
 import com.nexters.keyme.domain.notification.dto.ProblemSolvedNotificationRequest;
 import com.nexters.keyme.domain.notification.application.NotificationService;
